replace references to renamed conditional_selection_rate
